update HealthInsuranceAvgearn + pensionAvgEarn infra

+++ nts.uk/ctx/pr/core/nts.uk.ctx.pr.core.infra/src/main/java/nts/uk/ctx/pr/core/infra/entity/insurance/social/pensionavgearn/QismtPensionAvgearn.java
@@ -104,7 +104,7 @@ public class QismtPensionAvgearn implements Serializable {
 	/** The c pension male mny. */
 	@Basic(optional = false)
 	@Column(name = "C_PENSION_MALE_MNY")
-	private long cPensionMaleMny;
+	private BigDecimal cPensionMaleMny;
 
 	/** The c pension fem mny. */
 	@Basic(optional = false)
